+++ src/routes/index.ts
@@ -1,4 +1,5 @@
 import VueRouter from 'vue-router'
+import { generateUrl } from '@nextcloud/router'
 
 import ExplorePanel from '../components/routes/Explore.vue'
 import StarredPanel from '../components/routes/Starred.vue'
@@ -46,13 +47,7 @@ const getInitialRoute = function() {
 }
 
 const routes = [
-	// using
-	// { path: '/collections/all', component: CollectionGeneral, alias: '/' },
-	// instead of
 	{ path: '/', redirect: getInitialRoute() },
-	// would also be an option, but it currently does not work
-	// reliably with router-link due to
-	// https://github.com/vuejs/vue-router/issues/419
 	{
 		name: ROUTES.EXPLORE,
 		path: '/explore',
@@ -92,6 +87,8 @@ const routes = [
 ]
 
 export default new VueRouter({
+	mode: 'history',
+	base: generateUrl('/apps/news'),
 	linkActiveClass: 'active',
-	routes, // short for `routes: routes`
+	routes,
 })