+++ tests/javascript/unit/components/routes/Explore.spec.ts
@@ -1,19 +1,18 @@
 import axios from '@nextcloud/axios'
-import { shallowMount, createLocalVue } from '@vue/test-utils'
+import { shallowMount } from '@vue/test-utils'
+import { describe, expect, it, vi } from 'vitest'
 
 import Explore from '../../../../../src/components/routes/Explore.vue'
 
-jest.mock('@nextcloud/axios')
+vi.mock('@nextcloud/axios')
 
 describe('Explore.vue', () => {
 	'use strict'
-	const localVue = createLocalVue()
 
 	it('should initialize without showing AddFeed Component', () => {
 		(axios as any).get.mockResolvedValue({ data: { } });
 
 		const wrapper = shallowMount(Explore, {
-			localVue,
 			mocks: {
 				$store: {
 					state: {