+++ tests/javascript/unit/setup.ts
@@ -1,4 +1,5 @@
 import { config } from '@vue/test-utils'
+import { describe, it, expect, beforeAll, afterAll } from 'vitest'
 
 (global as any).appName = 'news'
 
@@ -7,16 +8,16 @@ import { OC } from './OC'
 (global as any).OC = new OC()
 
 // Mock nextcloud translate functions
-config.mocks.$t = function(_app: any, string: any) {
+config.global.mocks.$t = function(_app: any, string: any) {
 	return string
 }
-config.mocks.t = config.mocks.$t
-global.t = config.mocks.$t
+config.global.mocks.t = config.global.mocks.$t
+global.t = config.global.mocks.$t
 
-config.mocks.$n = function(app: any, singular: any) {
+config.global.mocks.$n = function(app: any, singular: any) {
 	return singular
 }
-config.mocks.n = config.mocks.$n
+config.global.mocks.n = config.global.mocks.$n
 
 afterAll(() => {
 	// TODO: afterAll tests?