+++ CHANGELOG.md
@@ -11,6 +11,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 
 ### Fixed
 - Wrong or none items in list after switching feed or folder
+- opml import/export not working after vue router changes
 
 
 # Releases